+++ poem/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.0.36] 2021-12-01
+
+- Add helper methods `Request::deserialize_path` and `Request::deserialize_query`.
+- Rename `error::ErrorInvalidPathParams` to `error::ParsePathError`.
+
 # [1.0.34] 2021-12-01
 
 - Implement `FromRequest` for `LocalAddr`.